automatically upload example datasets. Chao to make more changes to it.

+++ src/main/java/edu/uiuc/zenvisage/service/ZvMain.java
@@ -129,6 +129,11 @@ public class ZvMain {
 
 		UploadHandleServlet uploadHandler = new UploadHandleServlet();
 		List<String> names = uploadHandler.upload(request, response);
+		uploadDatasettoDB(names,true);
+	}
+
+		
+   public static void uploadDatasettoDB(List<String> names, boolean overwrite) throws SQLException, IOException{
 		SchemeToMetatable schemeToMetatable = new SchemeToMetatable();
 		
 		if (names.size() == 3) {
@@ -158,7 +163,7 @@ public class ZvMain {
 				sqlQueryExecutor.createTable(schemeToMetatable.createTableSQL);
 				sqlQueryExecutor.insertTable(names.get(0), names.get(1), schemeToMetatable.columns);
 				System.out.println(names.get(0) + " not exists! Created " + names.get(0) + " from "+names.get(1));
-			} else {//
+			} else if(overwrite) {//
 				sqlQueryExecutor.dropTable(names.get(0));
 				sqlQueryExecutor.createTable(schemeToMetatable.schemeFileToCreatTableSQL(names.get(2), names.get(0)));
 				sqlQueryExecutor.insertTable(names.get(0), names.get(1), schemeToMetatable.columns);
@@ -171,8 +176,12 @@ public class ZvMain {
 
 //			inMemoryDatabases.put(names.get(0), inMemoryDatabase);
 		}
+		
 	}
-
+	
+	
+	
+	
 //   public String runZQLCompleteQuery(String zqlQuery) throws IOException, InterruptedException, SQLException{
 //		  System.out.println(zqlQuery);
 //	   	  inMemoryDatabase = inMemoryDatabases.get("real_estate");
